+++ packages/form/src/components/Switch/index.tsx
@@ -20,9 +20,9 @@ const ProFormSwitch: React.FC<ProFormSwitchProps> = React.forwardRef(
         valueType="switch"
         mode="edit"
         fieldProps={{
-          ...fieldProps,
           unCheckedChildren,
           checkedChildren,
+          ...fieldProps,
         }}
         text={fieldProps?.checked}
         ref={ref}