+++ packages/layout/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.16.4](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.16.3...@ant-design/pro-layout@6.16.4) (2021-04-19)
+
+**Note:** Version bump only for package @ant-design/pro-layout
+
 ## [6.16.3](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.16.2...@ant-design/pro-layout@6.16.3) (2021-04-15)
 
 ### Bug Fixes